refactor: plonk solidity generate match expected gnark interface
