+++ backend/plonk/bn254/solidity.go
@@ -1,907 +1,951 @@
 package plonk
 
-const solidityTemplate = `
-// Warning this code was contributed into gnark here: 
-// https://github.com/ConsenSys/gnark/pull/358
-// 
-// It has not been audited and is provided as-is, we make no guarantees or warranties to its safety and reliability. 
-// 
-// According to https://eprint.iacr.org/archive/2019/953/1585767119.pdf
-pragma solidity ^0.8.0;
+const tmplSolidityVerifier = `pragma solidity ^0.8.0;
+
 pragma experimental ABIEncoderV2;
 
-library PairingsBn254 {
-    uint256 constant q_mod = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
+
+library Utils {
+
     uint256 constant r_mod = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
-    uint256 constant bn254_b_coeff = 3;
 
-    struct G1Point {
-        uint256 X;
-        uint256 Y;
-    }
-
-    struct Fr {
-        uint256 value;
-    }
-
-    function new_fr(uint256 fr) internal pure returns (Fr memory) {
-        require(fr < r_mod);
-        return Fr({value: fr});
-    }
-
-    function copy(Fr memory self) internal pure returns (Fr memory n) {
-        n.value = self.value;
-    }
-
-    function assign(Fr memory self, Fr memory other) internal pure {
-        self.value = other.value;
-    }
-
-    function inverse(Fr memory fr) internal view returns (Fr memory) {
-        require(fr.value != 0);
-        return pow(fr, r_mod-2);
-    }
-
-    function add_assign(Fr memory self, Fr memory other) internal pure {
-        self.value = addmod(self.value, other.value, r_mod);
-    }
-
-    function sub_assign(Fr memory self, Fr memory other) internal pure {
-        self.value = addmod(self.value, r_mod - other.value, r_mod);
-    }
-
-    function mul_assign(Fr memory self, Fr memory other) internal pure {
-        self.value = mulmod(self.value, other.value, r_mod);
-    }
-
-    function pow(Fr memory self, uint256 power) internal view returns (Fr memory) {
-        uint256[6] memory input = [32, 32, 32, self.value, power, r_mod];
-        uint256[1] memory result;
-        bool success;
-        assembly {
-            success := staticcall(gas(), 0x05, input, 0xc0, result, 0x20)
-        }
-        require(success);
-        return Fr({value: result[0]});
-    }
-
-    // Encoding of field elements is: X[0] * z + X[1]
-    struct G2Point {
-        uint[2] X;
-        uint[2] Y;
-    }
-
-    function P1() internal pure returns (G1Point memory) {
-        return G1Point(1, 2);
-    }
-
-    function new_g1(uint256 x, uint256 y) internal pure returns (G1Point memory) {
-        return G1Point(x, y);
-    }
-
-    function new_g1_checked(uint256 x, uint256 y) internal pure returns (G1Point memory) {
-        if (x == 0 && y == 0) {
-            // point of infinity is (0,0)
-            return G1Point(x, y);
-        }
-
-        // check encoding
-        require(x < q_mod);
-        require(y < q_mod);
-        // check on curve
-        uint256 lhs = mulmod(y, y, q_mod); // y^2
-        uint256 rhs = mulmod(x, x, q_mod); // x^2
-        rhs = mulmod(rhs, x, q_mod); // x^3
-        rhs = addmod(rhs, bn254_b_coeff, q_mod); // x^3 + b
-        require(lhs == rhs);
-
-        return G1Point(x, y);
-    }
-
-    function new_g2(uint256[2] memory x, uint256[2] memory y) internal pure returns (G2Point memory) {
-        return G2Point(x, y);
-    }
-
-    function copy_g1(G1Point memory self) internal pure returns (G1Point memory result) {
-        result.X = self.X;
-        result.Y = self.Y;
-    }
-
-    function P2() internal pure returns (G2Point memory) {
-        // for some reason ethereum expects to have c1*v + c0 form
-
-        return G2Point(
-            [0x198e9393920d483a7260bfb731fb5d25f1aa493335a9e71297e485b7aef312c2,
-            0x1800deef121f1e76426a00665e5c4479674322d4f75edadd46debd5cd992f6ed],
-            [0x090689d0585ff075ec9e99ad690c3395bc4b313370b38ef355acdadcd122975b,
-            0x12c85ea5db8c6deb4aab71808dcb408fe3d1e7690c43d37b4ce6cc0166fa7daa]
-        );
-    }
-
-    function negate(G1Point memory self) internal pure {
-        // The prime q in the base field F_q for G1
-        if (self.Y == 0) {
-            require(self.X == 0);
-            return;
-        }
-
-        self.Y = q_mod - self.Y;
-    }
-
-    function point_add(G1Point memory p1, G1Point memory p2)
-    internal view returns (G1Point memory r)
-    {
-        point_add_into_dest(p1, p2, r);
-        return r;
-    }
-
-    function point_add_assign(G1Point memory p1, G1Point memory p2)
-    internal view
-    {
-        point_add_into_dest(p1, p2, p1);
-    }
-
-    function point_add_into_dest(G1Point memory p1, G1Point memory p2, G1Point memory dest)
-    internal view
-    {
-        if (p2.X == 0 && p2.Y == 0) {
-            // we add zero, nothing happens
-            dest.X = p1.X;
-            dest.Y = p1.Y;
-            return;
-        } else if (p1.X == 0 && p1.Y == 0) {
-            // we add into zero, and we add non-zero point
-            dest.X = p2.X;
-            dest.Y = p2.Y;
-            return;
-        } else {
-            uint256[4] memory input;
-
-            input[0] = p1.X;
-            input[1] = p1.Y;
-            input[2] = p2.X;
-            input[3] = p2.Y;
-
-            bool success = false;
-            assembly {
-                success := staticcall(gas(), 6, input, 0x80, dest, 0x40)
-            }
-            require(success);
-        }
-    }
-
-    function point_sub_assign(G1Point memory p1, G1Point memory p2)
-    internal view
-    {
-        point_sub_into_dest(p1, p2, p1);
-    }
-
-    function point_sub_into_dest(G1Point memory p1, G1Point memory p2, G1Point memory dest)
-    internal view
-    {
-        if (p2.X == 0 && p2.Y == 0) {
-            // we subtracted zero, nothing happens
-            dest.X = p1.X;
-            dest.Y = p1.Y;
-            return;
-        } else if (p1.X == 0 && p1.Y == 0) {
-            // we subtract from zero, and we subtract non-zero point
-            dest.X = p2.X;
-            dest.Y = q_mod - p2.Y;
-            return;
-        } else {
-            uint256[4] memory input;
-
-            input[0] = p1.X;
-            input[1] = p1.Y;
-            input[2] = p2.X;
-            input[3] = q_mod - p2.Y;
-
-            bool success = false;
-            assembly {
-                success := staticcall(gas(), 6, input, 0x80, dest, 0x40)
-            }
-            require(success);
-        }
-    }
-
-    function point_mul(G1Point memory p, Fr memory s)
-    internal view returns (G1Point memory r)
-    {
-        point_mul_into_dest(p, s, r);
-        return r;
-    }
-
-    function point_mul_assign(G1Point memory p, Fr memory s)
-    internal view
-    {
-        point_mul_into_dest(p, s, p);
-    }
-
-    function point_mul_into_dest(G1Point memory p, Fr memory s, G1Point memory dest)
-    internal view
-    {
-        uint[3] memory input;
-        input[0] = p.X;
-        input[1] = p.Y;
-        input[2] = s.value;
-        bool success;
-        assembly {
-            success := staticcall(gas(), 7, input, 0x60, dest, 0x40)
-        }
-        require(success);
-    }
-
-    function pairing(G1Point[] memory p1, G2Point[] memory p2)
-    internal view returns (bool)
-    {
-        require(p1.length == p2.length);
-        uint elements = p1.length;
-        uint inputSize = elements * 6;
-        uint[] memory input = new uint[](inputSize);
-        for (uint i = 0; i < elements; i++)
-        {
-            input[i * 6 + 0] = p1[i].X;
-            input[i * 6 + 1] = p1[i].Y;
-            input[i * 6 + 2] = p2[i].X[0];
-            input[i * 6 + 3] = p2[i].X[1];
-            input[i * 6 + 4] = p2[i].Y[0];
-            input[i * 6 + 5] = p2[i].Y[1];
-        }
-        uint[1] memory out;
-        bool success;
-        assembly {
-            success := staticcall(gas(), 8, add(input, 0x20), mul(inputSize, 0x20), out, 0x20)
-        }
-        require(success);
-        return out[0] != 0;
-    }
-
-    /// Convenience method for a pairing check for two pairs.
-    function pairingProd2(G1Point memory a1, G2Point memory a2, G1Point memory b1, G2Point memory b2)
-    internal view returns (bool)
-    {
-        G1Point[] memory p1 = new G1Point[](2);
-        G2Point[] memory p2 = new G2Point[](2);
-        p1[0] = a1;
-        p1[1] = b1;
-        p2[0] = a2;
-        p2[1] = b2;
-        return pairing(p1, p2);
-    }
-}
-
-library TranscriptLibrary {
-    uint32 constant DST_0 = 0;
-    uint32 constant DST_1 = 1;
-    uint32 constant DST_CHALLENGE = 2;
-
-    struct Transcript {
-        bytes32 previous_randomness;
-        bytes bindings;
-        string name;
-        uint32 challenge_counter;
-    }
-
-    function new_transcript() internal pure returns (Transcript memory t) {
-        t.challenge_counter = 0;
-    }
-
-    function set_challenge_name(Transcript memory self, string memory name) internal pure {
-        self.name = name;
-    }
-
-    function update_with_u256(Transcript memory self, uint256 value) internal pure {
-        self.bindings = abi.encodePacked(self.bindings, value);
-    }
-
-    function update_with_fr(Transcript memory self, PairingsBn254.Fr memory value) internal pure {
-        self.bindings = abi.encodePacked(self.bindings, value.value);
-    }
-
-    function update_with_g1(Transcript memory self, PairingsBn254.G1Point memory p) internal pure {
-        self.bindings = abi.encodePacked(self.bindings, p.X, p.Y);
-    }
-
-    function get_encode(Transcript memory self) internal pure returns(bytes memory query) {
-        if (self.challenge_counter != 0) {
-            query = abi.encodePacked(self.name, self.previous_randomness, self.bindings);
-        } else {
-            query = abi.encodePacked(self.name, self.bindings);
-        }
-        return query;
-    }
-    function get_challenge(Transcript memory self) internal pure returns(PairingsBn254.Fr memory challenge) {
-        bytes32 query;
-        if (self.challenge_counter != 0) {
-            query = sha256(abi.encodePacked(self.name, self.previous_randomness, self.bindings));
-        } else {
-            query = sha256(abi.encodePacked(self.name, self.bindings));
-        }
-        self.challenge_counter += 1;
-        self.previous_randomness = query;
-        challenge = PairingsBn254.Fr({value: uint256(query) % PairingsBn254.r_mod});
-        self.bindings = "";
-    }
-}
-
-contract PlonkVerifier {
-    using PairingsBn254 for PairingsBn254.G1Point;
-    using PairingsBn254 for PairingsBn254.G2Point;
-    using PairingsBn254 for PairingsBn254.Fr;
-
-    using TranscriptLibrary for TranscriptLibrary.Transcript;
-
-    uint256 constant STATE_WIDTH = 3;
-
-    struct VerificationKey {
-        uint256 domain_size;
-        uint256 num_inputs;
-        PairingsBn254.Fr omega;                                     // w
-        PairingsBn254.G1Point[STATE_WIDTH+2] selector_commitments;  // STATE_WIDTH for witness + multiplication + constant
-        PairingsBn254.G1Point[STATE_WIDTH] permutation_commitments; // [Sσ1(x)],[Sσ2(x)],[Sσ3(x)]
-        PairingsBn254.Fr[STATE_WIDTH-1] permutation_non_residues;   // k1, k2
-        PairingsBn254.G2Point g2_x;
-    }
-
-    struct Proof {
-        uint256[] input_values;
-        PairingsBn254.G1Point[STATE_WIDTH] wire_commitments;  // [a(x)]/[b(x)]/[c(x)]
-        PairingsBn254.G1Point grand_product_commitment;      // [z(x)]
-        PairingsBn254.G1Point[STATE_WIDTH] quotient_poly_commitments;  // [t_lo]/[t_mid]/[t_hi]
-        PairingsBn254.Fr[STATE_WIDTH] wire_values_at_zeta;   // a(zeta)/b(zeta)/c(zeta)
-        PairingsBn254.Fr grand_product_at_zeta_omega;        // z(w*zeta)
-        PairingsBn254.Fr quotient_polynomial_at_zeta;        // t(zeta)
-        PairingsBn254.Fr linearization_polynomial_at_zeta;   // r(zeta)
-        PairingsBn254.Fr[STATE_WIDTH-1] permutation_polynomials_at_zeta;  // Sσ1(zeta),Sσ2(zeta)
-
-        PairingsBn254.G1Point opening_at_zeta_proof;            // [Wzeta]
-        PairingsBn254.G1Point opening_at_zeta_omega_proof;      // [Wzeta*omega]
-    }
-
-    struct PartialVerifierState {
-        PairingsBn254.Fr alpha;
-        PairingsBn254.Fr beta;
-        PairingsBn254.Fr gamma;
-        PairingsBn254.Fr v;
-        PairingsBn254.Fr u;
-        PairingsBn254.Fr zeta;
-        PairingsBn254.Fr[] cached_lagrange_evals;
-
-        PairingsBn254.G1Point cached_fold_quotient_ploy_commitments;
-    }
-
-    function verify_initial(
-		PartialVerifierState memory state,
-        Proof memory proof,
-        VerificationKey memory vk) internal view returns (bool) {
-
-        require(proof.input_values.length == vk.num_inputs, "not match");
-        require(vk.num_inputs >= 1, "inv input");
-        
-        TranscriptLibrary.Transcript memory t = TranscriptLibrary.new_transcript();
-        t.set_challenge_name("gamma");
-        for (uint256 i = 0; i < vk.permutation_commitments.length; i++) {
-            t.update_with_g1(vk.permutation_commitments[i]);
-        }
-        // this is gnark order: Ql, Qr, Qm, Qo, Qk
-        //
-        t.update_with_g1(vk.selector_commitments[0]);
-        t.update_with_g1(vk.selector_commitments[1]);
-        t.update_with_g1(vk.selector_commitments[3]);
-        t.update_with_g1(vk.selector_commitments[2]);
-        t.update_with_g1(vk.selector_commitments[4]);
-
-        for (uint256 i = 0; i < proof.input_values.length; i++) {
-            t.update_with_u256(proof.input_values[i]);
-        }
-        state.gamma = t.get_challenge();
-
-        t.set_challenge_name("beta");
-        state.beta = t.get_challenge();
-
-        t.set_challenge_name("alpha");
-        t.update_with_g1(proof.grand_product_commitment);
-        state.alpha = t.get_challenge();
-
-        t.set_challenge_name("zeta");
-        for (uint256 i = 0; i < proof.quotient_poly_commitments.length; i++) {
-            t.update_with_g1(proof.quotient_poly_commitments[i]);
-        }
-        state.zeta = t.get_challenge();
-
-        uint256[] memory lagrange_poly_numbers = new uint256[](vk.num_inputs);
-        for (uint256 i = 0; i < lagrange_poly_numbers.length; i++) {
-            lagrange_poly_numbers[i] = i;
-        }
-        state.cached_lagrange_evals = batch_evaluate_lagrange_poly_out_of_domain(
-            lagrange_poly_numbers,
-            vk.domain_size,
-            vk.omega, state.zeta
-        );
-
-        bool valid = verify_quotient_poly_eval_at_zeta(state, proof, vk);
-        return valid;
-    }
-
-    function verify_commitments(
-        PartialVerifierState memory state,
-        Proof memory proof,
-        VerificationKey memory vk
-    ) internal view returns (bool) {
-        PairingsBn254.G1Point memory d = reconstruct_d(state, proof, vk);
-
-        PairingsBn254.G1Point memory tmp_g1 = PairingsBn254.P1();
-
-        PairingsBn254.Fr memory aggregation_challenge = PairingsBn254.new_fr(1);
-
-        PairingsBn254.G1Point memory commitment_aggregation = PairingsBn254.copy_g1(state.cached_fold_quotient_ploy_commitments);
-        PairingsBn254.Fr memory tmp_fr = PairingsBn254.new_fr(1);
-
-        aggregation_challenge.mul_assign(state.v);
-        commitment_aggregation.point_add_assign(d);
-
-        for (uint i = 0; i < proof.wire_commitments.length; i++) {
-            aggregation_challenge.mul_assign(state.v);
-            tmp_g1 = proof.wire_commitments[i].point_mul(aggregation_challenge);
-            commitment_aggregation.point_add_assign(tmp_g1);
-        }
-
-        for (uint i = 0; i < vk.permutation_commitments.length - 1; i++) {
-            aggregation_challenge.mul_assign(state.v);
-            tmp_g1 = vk.permutation_commitments[i].point_mul(aggregation_challenge);
-            commitment_aggregation.point_add_assign(tmp_g1);
-        }
-
-        // collect opening values
-        aggregation_challenge = PairingsBn254.new_fr(1);
-
-        PairingsBn254.Fr memory aggregated_value = PairingsBn254.copy(proof.quotient_polynomial_at_zeta);
-
-        aggregation_challenge.mul_assign(state.v);
-
-        tmp_fr.assign(proof.linearization_polynomial_at_zeta);
-        tmp_fr.mul_assign(aggregation_challenge);
-        aggregated_value.add_assign(tmp_fr);
-
-        for (uint i = 0; i < proof.wire_values_at_zeta.length; i++) {
-            aggregation_challenge.mul_assign(state.v);
-
-            tmp_fr.assign(proof.wire_values_at_zeta[i]);
-            tmp_fr.mul_assign(aggregation_challenge);
-            aggregated_value.add_assign(tmp_fr);
-        }
-
-        for (uint i = 0; i < proof.permutation_polynomials_at_zeta.length; i++) {
-            aggregation_challenge.mul_assign(state.v);
-
-            tmp_fr.assign(proof.permutation_polynomials_at_zeta[i]);
-            tmp_fr.mul_assign(aggregation_challenge);
-            aggregated_value.add_assign(tmp_fr);
-        }
-        tmp_fr.assign(proof.grand_product_at_zeta_omega);
-        tmp_fr.mul_assign(state.u);
-        aggregated_value.add_assign(tmp_fr);
-
-        commitment_aggregation.point_sub_assign(PairingsBn254.P1().point_mul(aggregated_value));
-
-        PairingsBn254.G1Point memory pair_with_generator = commitment_aggregation;
-        pair_with_generator.point_add_assign(proof.opening_at_zeta_proof.point_mul(state.zeta));
-
-        tmp_fr.assign(state.zeta);
-        tmp_fr.mul_assign(vk.omega);
-        tmp_fr.mul_assign(state.u);
-        pair_with_generator.point_add_assign(proof.opening_at_zeta_omega_proof.point_mul(tmp_fr));
-
-        PairingsBn254.G1Point memory pair_with_x = proof.opening_at_zeta_omega_proof.point_mul(state.u);
-        pair_with_x.point_add_assign(proof.opening_at_zeta_proof);
-        pair_with_x.negate();
-
-        return PairingsBn254.pairingProd2(pair_with_generator, PairingsBn254.P2(), pair_with_x, vk.g2_x);
-    }
-
-    function reconstruct_d(
-        PartialVerifierState memory state,
-        Proof memory proof,
-        VerificationKey memory vk
-    ) internal view returns (PairingsBn254.G1Point memory res) {
-        res = PairingsBn254.copy_g1(vk.selector_commitments[STATE_WIDTH + 1]);
-
-        PairingsBn254.G1Point memory tmp_g1 = PairingsBn254.P1();
-        PairingsBn254.Fr memory tmp_fr = PairingsBn254.new_fr(0);
-
-        // addition gates
-        for (uint256 i = 0; i < STATE_WIDTH; i++) {
-            tmp_g1 = vk.selector_commitments[i].point_mul(proof.wire_values_at_zeta[i]);
-            res.point_add_assign(tmp_g1);
-        }
-
-        // multiplication gate
-        tmp_fr.assign(proof.wire_values_at_zeta[0]);
-        tmp_fr.mul_assign(proof.wire_values_at_zeta[1]);
-        tmp_g1 = vk.selector_commitments[STATE_WIDTH].point_mul(tmp_fr);
-        res.point_add_assign(tmp_g1);
-
-        // z * non_res * beta + gamma + a
-        PairingsBn254.Fr memory grand_product_part_at_z = PairingsBn254.copy(state.zeta);
-        grand_product_part_at_z.mul_assign(state.beta);
-        grand_product_part_at_z.add_assign(proof.wire_values_at_zeta[0]);
-        grand_product_part_at_z.add_assign(state.gamma);
-        for (uint256 i = 0; i < vk.permutation_non_residues.length; i++) {
-            tmp_fr.assign(state.zeta);
-            tmp_fr.mul_assign(vk.permutation_non_residues[i]);
-            tmp_fr.mul_assign(state.beta);
-            tmp_fr.add_assign(state.gamma);
-            tmp_fr.add_assign(proof.wire_values_at_zeta[i+1]);
-
-            grand_product_part_at_z.mul_assign(tmp_fr);
-        }
-
-        grand_product_part_at_z.mul_assign(state.alpha);
-
-        tmp_fr.assign(state.cached_lagrange_evals[0]);
-        tmp_fr.mul_assign(state.alpha);
-        tmp_fr.mul_assign(state.alpha);
-        // NOTICE
-        grand_product_part_at_z.sub_assign(tmp_fr);
-        PairingsBn254.Fr memory last_permutation_part_at_z = PairingsBn254.new_fr(1);
-        for (uint256 i = 0; i < proof.permutation_polynomials_at_zeta.length; i++) {
-            tmp_fr.assign(state.beta);
-            tmp_fr.mul_assign(proof.permutation_polynomials_at_zeta[i]);
-            tmp_fr.add_assign(state.gamma);
-            tmp_fr.add_assign(proof.wire_values_at_zeta[i]);
-
-            last_permutation_part_at_z.mul_assign(tmp_fr);
-        }
-
-        last_permutation_part_at_z.mul_assign(state.beta);
-        last_permutation_part_at_z.mul_assign(proof.grand_product_at_zeta_omega);
-        last_permutation_part_at_z.mul_assign(state.alpha);
-
-        // gnark implementation: add third part and sub second second part
-        // plonk paper implementation: add second part and sub third part
-        /*
-        tmp_g1 = proof.grand_product_commitment.point_mul(grand_product_part_at_z);
-        tmp_g1.point_sub_assign(vk.permutation_commitments[STATE_WIDTH - 1].point_mul(last_permutation_part_at_z));
-        */
-        // add to the linearization
-
-        tmp_g1 = vk.permutation_commitments[STATE_WIDTH - 1].point_mul(last_permutation_part_at_z);
-        tmp_g1.point_sub_assign(proof.grand_product_commitment.point_mul(grand_product_part_at_z));
-        res.point_add_assign(tmp_g1);
-
-        generate_uv_challenge(state, proof, vk, res);
-
-        res.point_mul_assign(state.v);
-        res.point_add_assign(proof.grand_product_commitment.point_mul(state.u));
-    }
-
-    // gnark v generation process:
-    // sha256(zeta, proof.quotient_poly_commitments, linearizedPolynomialDigest, proof.wire_commitments, vk.permutation_commitments[0..1], )
-    // NOTICE: gnark use "gamma" name for v, it's not reasonable
-    // NOTICE: gnark use zeta^(n+2) which is a bit different with plonk paper
-    // generate_v_challenge();
-    function generate_uv_challenge(
-        PartialVerifierState memory state,
-        Proof memory proof,
-        VerificationKey memory vk,
-        PairingsBn254.G1Point memory linearization_point) view internal {
-        TranscriptLibrary.Transcript memory transcript = TranscriptLibrary.new_transcript();
-        transcript.set_challenge_name("gamma");
-        transcript.update_with_fr(state.zeta);
-        PairingsBn254.Fr memory zeta_plus_two = PairingsBn254.copy(state.zeta);
-        PairingsBn254.Fr memory n_plus_two = PairingsBn254.new_fr(vk.domain_size);
-        n_plus_two.add_assign(PairingsBn254.new_fr(2));
-        zeta_plus_two = zeta_plus_two.pow(n_plus_two.value);
-        state.cached_fold_quotient_ploy_commitments = PairingsBn254.copy_g1(proof.quotient_poly_commitments[STATE_WIDTH-1]);
-        for (uint256 i = 0; i < STATE_WIDTH - 1; i++) {
-            state.cached_fold_quotient_ploy_commitments.point_mul_assign(zeta_plus_two);
-            state.cached_fold_quotient_ploy_commitments.point_add_assign(proof.quotient_poly_commitments[STATE_WIDTH - 2 - i]);
-        }
-        transcript.update_with_g1(state.cached_fold_quotient_ploy_commitments);
-        transcript.update_with_g1(linearization_point);
-
-        for (uint256 i = 0; i < proof.wire_commitments.length; i++) {
-            transcript.update_with_g1(proof.wire_commitments[i]);
-        }
-        for (uint256 i = 0; i < vk.permutation_commitments.length - 1; i++) {
-            transcript.update_with_g1(vk.permutation_commitments[i]);
-        }
-        state.v = transcript.get_challenge();
-        // gnark use local randomness to generate u
-        // we use opening_at_zeta_proof and opening_at_zeta_omega_proof
-        transcript.set_challenge_name("u");
-        transcript.update_with_g1(proof.opening_at_zeta_proof);
-        transcript.update_with_g1(proof.opening_at_zeta_omega_proof);
-        state.u = transcript.get_challenge();
-    }
-
-    function batch_evaluate_lagrange_poly_out_of_domain(
-        uint256[] memory poly_nums,
-        uint256 domain_size,
-        PairingsBn254.Fr memory omega,
-        PairingsBn254.Fr memory at
-    ) internal view returns (PairingsBn254.Fr[] memory res) {
-        PairingsBn254.Fr memory one = PairingsBn254.new_fr(1);
-        PairingsBn254.Fr memory tmp_1 = PairingsBn254.new_fr(0);
-        PairingsBn254.Fr memory tmp_2 = PairingsBn254.new_fr(domain_size);
-        PairingsBn254.Fr memory vanishing_at_zeta = at.pow(domain_size);
-        vanishing_at_zeta.sub_assign(one);
-        // we can not have random point z be in domain
-        require(vanishing_at_zeta.value != 0);
-        PairingsBn254.Fr[] memory nums = new PairingsBn254.Fr[](poly_nums.length);
-        PairingsBn254.Fr[] memory dens = new PairingsBn254.Fr[](poly_nums.length);
-        // numerators in a form omega^i * (z^n - 1)
-        // denoms in a form (z - omega^i) * N
-        for (uint i = 0; i < poly_nums.length; i++) {
-            tmp_1 = omega.pow(poly_nums[i]); // power of omega
-            nums[i].assign(vanishing_at_zeta);
-            nums[i].mul_assign(tmp_1);
-
-            dens[i].assign(at); // (X - omega^i) * N
-            dens[i].sub_assign(tmp_1);
-            dens[i].mul_assign(tmp_2); // mul by domain size
-        }
-
-        PairingsBn254.Fr[] memory partial_products = new PairingsBn254.Fr[](poly_nums.length);
-        partial_products[0].assign(PairingsBn254.new_fr(1));
-        for (uint i = 1; i < dens.length; i++) {
-            partial_products[i].assign(dens[i-1]);
-            partial_products[i].mul_assign(partial_products[i-1]);
-        }
-
-        tmp_2.assign(partial_products[partial_products.length - 1]);
-        tmp_2.mul_assign(dens[dens.length - 1]);
-        tmp_2 = tmp_2.inverse(); // tmp_2 contains a^-1 * b^-1 (with! the last one)
-
-        for (uint i = dens.length; i > 0; i--) {
-            tmp_1.assign(tmp_2); // all inversed
-            tmp_1.mul_assign(partial_products[i-1]); // clear lowest terms
-            tmp_2.mul_assign(dens[i-1]);
-            dens[i-1].assign(tmp_1);
-        }
-
-        for (uint i = 0; i < nums.length; i++) {
-            nums[i].mul_assign(dens[i]);
-        }
-
-        return nums;
-    }
-
-    // plonk paper verify process step8: Compute quotient polynomial evaluation
-    function verify_quotient_poly_eval_at_zeta(
-        PartialVerifierState memory state,
-        Proof memory proof,
-        VerificationKey memory vk
-    ) internal view returns (bool) {
-        PairingsBn254.Fr memory lhs = evaluate_vanishing(vk.domain_size, state.zeta);
-        require(lhs.value != 0); // we can not check a polynomial relationship if point z is in the domain
-        lhs.mul_assign(proof.quotient_polynomial_at_zeta);
-
-        PairingsBn254.Fr memory quotient_challenge = PairingsBn254.new_fr(1);
-        PairingsBn254.Fr memory rhs = PairingsBn254.copy(proof.linearization_polynomial_at_zeta);
-
-        // public inputs
-        PairingsBn254.Fr memory tmp = PairingsBn254.new_fr(0);
-        for (uint256 i = 0; i < proof.input_values.length; i++) {
-            tmp.assign(state.cached_lagrange_evals[i]);
-            tmp.mul_assign(PairingsBn254.new_fr(proof.input_values[i]));
-            rhs.add_assign(tmp);
-        }
-
-        quotient_challenge.mul_assign(state.alpha);
-
-        PairingsBn254.Fr memory z_part = PairingsBn254.copy(proof.grand_product_at_zeta_omega);
-        for (uint256 i = 0; i < proof.permutation_polynomials_at_zeta.length; i++) {
-            tmp.assign(proof.permutation_polynomials_at_zeta[i]);
-            tmp.mul_assign(state.beta);
-            tmp.add_assign(state.gamma);
-            tmp.add_assign(proof.wire_values_at_zeta[i]);
-
-            z_part.mul_assign(tmp);
-        }
-
-        tmp.assign(state.gamma);
-        // we need a wire value of the last polynomial in enumeration
-        tmp.add_assign(proof.wire_values_at_zeta[STATE_WIDTH - 1]);
-
-        z_part.mul_assign(tmp);
-        z_part.mul_assign(quotient_challenge);
-
-        // NOTICE: this is different with plonk paper
-        // plonk paper should be: rhs.sub_assign(z_part);
-        rhs.add_assign(z_part);
-
-        quotient_challenge.mul_assign(state.alpha);
-
-        tmp.assign(state.cached_lagrange_evals[0]);
-        tmp.mul_assign(quotient_challenge);
-
-        rhs.sub_assign(tmp);
-
-        return lhs.value == rhs.value;
-    }
-
-    function evaluate_vanishing(
-        uint256 domain_size,
-        PairingsBn254.Fr memory at
-    ) internal view returns (PairingsBn254.Fr memory res) {
-        res = at.pow(domain_size);
-        res.sub_assign(PairingsBn254.new_fr(1));
-    }
-
-	// This verifier is for a PLONK with a state width 3
-    // and main gate equation
-    // q_a(X) * a(X) + 
-    // q_b(X) * b(X) + 
-    // q_c(X) * c(X) +
-    // q_m(X) * a(X) * b(X) + 
-    // q_constants(X)+
-    // where q_{}(X) are selectors a, b, c - state (witness) polynomials
+    /**
+    * @dev ExpandMsgXmd expands msg to a slice of lenInBytes bytes.
+    *      https://tools.ietf.org/html/draft-irtf-cfrg-hash-to-curve-06#section-5
+    *      https://tools.ietf.org/html/rfc8017#section-4.1 (I2OSP/O2ISP)
+    */
+    function expand_msg(uint256 x, uint256 y) public pure returns(uint8[48] memory res){
     
-    function verify(Proof memory proof, VerificationKey memory vk) internal view returns (bool) {
-        PartialVerifierState memory state;
+        string memory dst = "BSB22-Plonk";
+
+        //uint8[64] memory pad; // 64 is sha256 block size.
+        // sha256(pad || msg || (0 || 48 || 0) || dst || 11)
+        bytes memory tmp;
+        uint8 zero = 0;
+        uint8 lenInBytes = 48;
+        uint8 sizeDomain = 11; // size of dst
         
-        bool valid = verify_initial(state, proof, vk);
-        
-        if (valid == false) {
-            return false;
+        for (uint i=0; i<64; i++){
+            tmp = abi.encodePacked(tmp, zero);
         }
-        
-        valid = verify_commitments(state, proof, vk);
-        
-        return valid;
+        tmp = abi.encodePacked(tmp, x, y, zero, lenInBytes, zero, dst, sizeDomain);
+        bytes32 b0 = sha256(tmp);
+
+        tmp = abi.encodePacked(b0, uint8(1), dst, sizeDomain);
+        bytes32 b1 = sha256(tmp);
+        for (uint i=0; i<32; i++){
+            res[i] = uint8(b1[i]);
+        }
+
+        tmp = abi.encodePacked(uint8(b0[0]) ^ uint8(b1[0]));
+        for (uint i=1; i<32; i++){
+            tmp = abi.encodePacked(tmp, uint8(b0[i]) ^ uint8(b1[i]));
+        }
+
+        tmp = abi.encodePacked(tmp, uint8(2), dst, sizeDomain);
+        b1 = sha256(tmp);
+
+        // TODO handle the size of the dst (check gnark-crypto)
+        for (uint i=0; i<16; i++){
+            res[i+32] = uint8(b1[i]);
+        }
+
+        return res;
     }
+
+  /**
+   * @dev cf https://tools.ietf.org/html/draft-irtf-cfrg-hash-to-curve-06#section-5.2
+   * corresponds to https://github.com/ConsenSys/gnark-crypto/blob/develop/ecc/bn254/fr/element.go
+   */
+    function hash_fr(uint256 x, uint256 y) internal pure returns(uint256 res) {
+
+        // interpret a as a bigEndian integer and reduce it mod r
+        uint8[48] memory xmsg = expand_msg(x, y);
+        // uint8[48] memory xmsg = [0x44, 0x74, 0xb5, 0x29, 0xd7, 0xfb, 0x29, 0x88, 0x3a, 0x7a, 0xc1, 0x65, 0xfd, 0x72, 0xce, 0xd0, 0xd4, 0xd1, 0x3f, 0x9e, 0x85, 0x8a, 0x3, 0x86, 0x1c, 0x90, 0x83, 0x1e, 0x94, 0xdc, 0xfc, 0x1d, 0x70, 0x82, 0xf5, 0xbf, 0x30, 0x3, 0x39, 0x87, 0x21, 0x38, 0x15, 0xed, 0x12, 0x75, 0x44, 0x6a];
+
+        // reduce xmsg mod r, where xmsg is intrepreted in big endian 
+        // (as SetBytes does for golang's Big.Int library).
+        for (uint i=0; i<32; i++){
+            res += uint256(xmsg[47-i])<<(8*i);
+        }
+        res = res % r_mod;
+        uint256 tmp;
+        for (uint i=0; i<16; i++){
+            tmp += uint256(xmsg[15-i])<<(8*i);
+        }
+
+        // 2**256%r
+        uint256 b = 6350874878119819312338956282401532410528162663560392320966563075034087161851; 
+        assembly {
+            tmp := mulmod(tmp, b, r_mod)
+            res := addmod(res, tmp, r_mod)
+        }
+
+        return res;
+    }
+
 }
 
-contract KeyedPlonkVerifier is PlonkVerifier {
-    uint256 constant SERIALIZED_PROOF_LENGTH = 26;
-	using PairingsBn254 for PairingsBn254.Fr;
-    function get_verification_key() internal pure returns(VerificationKey memory vk) {
-        vk.domain_size = {{.Size}};
-        vk.num_inputs = {{.NbPublicVariables}};
-        vk.omega = PairingsBn254.new_fr(uint256({{.Generator.String}}));
-        vk.selector_commitments[0] = PairingsBn254.new_g1(
-        	uint256({{.Ql.X.String}}),
-        	uint256({{.Ql.Y.String}})
-        );
-        vk.selector_commitments[1] = PairingsBn254.new_g1(
-			uint256({{.Qr.X.String}}),
-			uint256({{.Qr.Y.String}})
-        );
-        vk.selector_commitments[2] = PairingsBn254.new_g1(
-			uint256({{.Qo.X.String}}),
-			uint256({{.Qo.Y.String}})
-        );
-        vk.selector_commitments[3] = PairingsBn254.new_g1(
-			uint256({{.Qm.X.String}}),
-			uint256({{.Qm.Y.String}})
-        );
-        vk.selector_commitments[4] = PairingsBn254.new_g1(
-			uint256({{.Qk.X.String}}),
-			uint256({{.Qk.Y.String}})
-        );
+library PlonkVerifier {
 
-        vk.permutation_commitments[0] = PairingsBn254.new_g1(
-        	uint256({{(index .S 0).X.String}}),
-			uint256({{(index .S 0).Y.String}})
-        );
-        vk.permutation_commitments[1] = PairingsBn254.new_g1(
-			uint256({{(index .S 1).X.String}}),
-			uint256({{(index .S 1).Y.String}})
-        );
-        vk.permutation_commitments[2] = PairingsBn254.new_g1(
-			uint256({{(index .S 2).X.String}}),
-			uint256({{(index .S 2).Y.String}})
-        );
+  using Utils for *;
+  uint256 constant r_mod = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
+  uint256 constant p_mod = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
+  {{ range $index, $element := .Kzg.G2 }}
+  uint256 constant g2_srs_{{ $index }}_x_0 = {{ (fpptr $element.X.A1).String }};
+  uint256 constant g2_srs_{{ $index }}_x_1 = {{ (fpptr $element.X.A0).String }};
+  uint256 constant g2_srs_{{ $index }}_y_0 = {{ (fpptr $element.Y.A1).String }};
+  uint256 constant g2_srs_{{ $index }}_y_1 = {{ (fpptr $element.Y.A0).String }};
+  {{ end }}
+  // ----------------------- vk ---------------------
+  uint256 constant vk_domain_size = {{ .Size }};
+  uint256 constant vk_inv_domain_size = {{ (frptr .SizeInv).String }};
+  uint256 constant vk_omega = {{ (frptr .Generator).String }};
+  uint256 constant vk_ql_com_x = {{ (fpptr .Ql.X).String }};
+  uint256 constant vk_ql_com_y = {{ (fpptr .Ql.Y).String }};
+  uint256 constant vk_qr_com_x = {{ (fpptr .Qr.X).String }};
+  uint256 constant vk_qr_com_y = {{ (fpptr .Qr.Y).String }};
+  uint256 constant vk_qm_com_x = {{ (fpptr .Qm.X).String }};
+  uint256 constant vk_qm_com_y = {{ (fpptr .Qm.Y).String }};
+  uint256 constant vk_qo_com_x = {{ (fpptr .Qo.X).String }};
+  uint256 constant vk_qo_com_y = {{ (fpptr .Qo.Y).String }};
+  uint256 constant vk_qk_com_x = {{ (fpptr .Qk.X).String }};
+  uint256 constant vk_qk_com_y = {{ (fpptr .Qk.Y).String }};
+  {{ range $index, $element := .S }}
+  uint256 constant vk_s{{ inc $index }}_com_x = {{ (fpptr $element.X).String }};
+  uint256 constant vk_s{{ inc $index }}_com_y = {{ (fpptr $element.Y).String }};
+  {{ end }}
+  uint256 constant vk_coset_shift = 5;
+  
+  {{ range $index, $element := .Qcp}}
+  uint256 constant vk_selector_commitments_commit_api_{{ $index }}_x = {{ (fpptr $element.X).String }};
+  uint256 constant vk_selector_commitments_commit_api_{{ $index }}_y = {{ (fpptr $element.Y).String }};
+  {{ end }}
 
-        vk.permutation_non_residues[0] = PairingsBn254.new_fr(
-        	uint256({{.CosetShift.String}})
-        );
-        vk.permutation_non_residues[1] = PairingsBn254.copy(
-			vk.permutation_non_residues[0]
-        );
-		vk.permutation_non_residues[1].mul_assign(vk.permutation_non_residues[0]);
+  {{ if (gt (len .CommitmentConstraintIndexes) 0 )}}
+  function load_vk_commitments_indices_commit_api(uint256[] memory v)
+  internal view {
+    assembly {
+    let _v := add(v, 0x20)
+    {{ range .CommitmentConstraintIndexes }}
+    mstore(_v, {{ . }})
+    _v := add(_v, 0x20)
+    {{ end }}
+    }
+  }
+  {{ end }}
+  uint256 constant vk_nb_commitments_commit_api = {{ len .CommitmentConstraintIndexes }};
 
-        vk.g2_x = PairingsBn254.new_g2(
-			[uint256({{(index .KZGSRS.G2 1).X.A1.String}}),
-			uint256({{(index .KZGSRS.G2 1).X.A0.String}})],
-			[uint256({{(index .KZGSRS.G2 1).Y.A1.String}}),
-			uint256({{(index .KZGSRS.G2 1).Y.A0.String}})]
-        );
+  // ------------------------------------------------
+
+  // offset proof
+  uint256 constant proof_l_com_x = 0x20;
+  uint256 constant proof_l_com_y = 0x40;
+  uint256 constant proof_r_com_x = 0x60;
+  uint256 constant proof_r_com_y = 0x80;
+  uint256 constant proof_o_com_x = 0xa0;
+  uint256 constant proof_o_com_y = 0xc0;
+
+  // h = h_0 + x^{n+2}h_1 + x^{2(n+2)}h_2
+  uint256 constant proof_h_0_x = 0xe0; 
+  uint256 constant proof_h_0_y = 0x100;
+  uint256 constant proof_h_1_x = 0x120;
+  uint256 constant proof_h_1_y = 0x140;
+  uint256 constant proof_h_2_x = 0x160;
+  uint256 constant proof_h_2_y = 0x180;
+
+  // wire values at zeta
+  uint256 constant proof_l_at_zeta = 0x1a0;
+  uint256 constant proof_r_at_zeta = 0x1c0;
+  uint256 constant proof_o_at_zeta = 0x1e0;
+
+  //uint256[STATE_WIDTH-1] permutation_polynomials_at_zeta; // Sσ1(zeta),Sσ2(zeta)
+  uint256 constant proof_s1_at_zeta = 0x200; // Sσ1(zeta)
+  uint256 constant proof_s2_at_zeta = 0x220; // Sσ2(zeta)
+
+  //Bn254.G1Point grand_product_commitment;                 // [z(x)]
+  uint256 constant proof_grand_product_commitment_x = 0x240;
+  uint256 constant proof_grand_product_commitment_y = 0x260;
+
+  uint256 constant proof_grand_product_at_zeta_omega = 0x280;                    // z(w*zeta)
+  uint256 constant proof_quotient_polynomial_at_zeta = 0x2a0;                    // t(zeta)
+  uint256 constant proof_linearised_polynomial_at_zeta = 0x2c0;               // r(zeta)
+
+  // Folded proof for the opening of H, linearised poly, l, r, o, s_1, s_2, qcp
+  uint256 constant proof_batch_opening_at_zeta_x = 0x2e0;            // [Wzeta]
+  uint256 constant proof_batch_opening_at_zeta_y = 0x300;
+
+  //Bn254.G1Point opening_at_zeta_omega_proof;      // [Wzeta*omega]
+  uint256 constant proof_opening_at_zeta_omega_x = 0x320;
+  uint256 constant proof_opening_at_zeta_omega_y = 0x340;
+  
+  uint256 constant proof_openings_selector_commit_api_at_zeta = 0x360;
+  // -> next part of proof is 
+  // [ openings_selector_commits || commitments_wires_commit_api]
+
+  // -------- offset state
+
+  // challenges to check the claimed quotient
+  uint256 constant state_alpha = 0x00;
+  uint256 constant state_beta = 0x20;
+  uint256 constant state_gamma = 0x40;
+  uint256 constant state_zeta = 0x60;
+
+  // challenges related to KZG
+  uint256 constant state_sv = 0x80;
+  uint256 constant state_su = 0xa0;
+
+  // reusable value
+  uint256 constant state_alpha_square_lagrange = 0xc0;
+
+  // commitment to H
+  // Bn254.G1Point folded_h;
+  uint256 constant state_folded_h_x = 0xe0;
+  uint256 constant state_folded_h_y = 0x100;
+
+  // commitment to the linearised polynomial
+  uint256 constant state_linearised_polynomial_x = 0x120;
+  uint256 constant state_linearised_polynomial_y = 0x140;
+
+  // Folded proof for the opening of H, linearised poly, l, r, o, s_1, s_2, qcp
+  // Kzg.OpeningProof folded_proof;
+  uint256 constant state_folded_claimed_values = 0x160;
+
+  // folded digests of H, linearised poly, l, r, o, s_1, s_2, qcp
+  // Bn254.G1Point folded_digests;
+  uint256 constant state_folded_digests_x = 0x180;
+  uint256 constant state_folded_digests_y = 0x1a0;
+
+  uint256 constant state_pi = 0x1c0;
+
+  uint256 constant state_zeta_power_n_minus_one = 0x1e0;
+  uint256 constant state_alpha_square_lagrange_one = 0x200;
+
+  uint256 constant state_gamma_kzg = 0x220;
+
+  uint256 constant state_success = 0x240;
+  uint256 constant state_check_var = 0x260; // /!\ this slot is used for debugging only
+
+
+  uint256 constant state_last_mem = 0x280;
+
+  event PrintUint256(uint256 a);
+
+  function derive_gamma_beta_alpha_zeta(bytes memory proof, uint256[] memory public_inputs)
+  internal returns(uint256, uint256, uint256, uint256) {
+
+    uint256 gamma;
+    uint256 beta;
+    uint256 alpha;
+    uint256 zeta;
+
+    assembly {
+
+      let mem := mload(0x40)
+
+      derive_gamma(proof, public_inputs)
+      gamma := mload(mem)
+
+      derive_beta(proof, gamma)
+      beta := mload(mem)
+
+      derive_alpha(proof, beta)
+      alpha := mload(mem)
+
+      derive_zeta(proof, alpha)
+      zeta := mload(mem)
+
+      gamma := mod(gamma, r_mod)
+      beta := mod(beta, r_mod)
+      alpha := mod(alpha, r_mod)
+      zeta := mod(zeta, r_mod)
+
+      function derive_gamma(aproof, pub_inputs) {
+        
+        let mPtr := mload(0x40)
+
+        // gamma
+        mstore(mPtr, 0x67616d6d61) // "gamma"
+
+        mstore(add(mPtr, 0x20), vk_s1_com_x)
+        mstore(add(mPtr, 0x40), vk_s1_com_y)
+        mstore(add(mPtr, 0x60), vk_s2_com_x)
+        mstore(add(mPtr, 0x80), vk_s2_com_y)
+        mstore(add(mPtr, 0xa0), vk_s3_com_x)
+        mstore(add(mPtr, 0xc0), vk_s3_com_y)
+        mstore(add(mPtr, 0xe0), vk_ql_com_x)
+        mstore(add(mPtr, 0x100), vk_ql_com_y)
+        mstore(add(mPtr, 0x120), vk_qr_com_x)
+        mstore(add(mPtr, 0x140), vk_qr_com_y)
+        mstore(add(mPtr, 0x160), vk_qm_com_x)
+        mstore(add(mPtr, 0x180), vk_qm_com_y)
+        mstore(add(mPtr, 0x1a0), vk_qo_com_x)
+        mstore(add(mPtr, 0x1c0), vk_qo_com_y)
+        mstore(add(mPtr, 0x1e0), vk_qk_com_x)
+        mstore(add(mPtr, 0x200), vk_qk_com_y)
+
+        let pi := add(pub_inputs, 0x20)
+        let _mPtr := add(mPtr, 0x220)
+        for {let i:=0} lt(i, mload(pub_inputs)) {i:=add(i,1)}
+        {
+          mstore(_mPtr, mload(pi))
+          pi := add(pi, 0x20)
+          _mPtr := add(_mPtr, 0x20)
+        }
+
+        let _proof := add(aproof, proof_openings_selector_commit_api_at_zeta)
+        _proof := add(_proof, mul(vk_nb_commitments_commit_api, 0x20))
+        for {let i:=0} lt(i, vk_nb_commitments_commit_api) {i:=add(i,1)}
+        {
+          mstore(_mPtr, mload(_proof))
+          mstore(add(_mPtr, 0x20), mload(add(_proof, 0x20)))
+          _mPtr := add(_mPtr, 0x40)
+          _proof := add(_proof, 0x40)
+        }
+        // pop(staticcall(sub(gas(), 2000), 0x2, add(mPtr, 0x1b), 0x2a5, mPtr, 0x20)) //0x1b -> 000.."gamma"
+
+        mstore(_mPtr, mload(add(aproof, proof_l_com_x)))
+        mstore(add(_mPtr, 0x20), mload(add(aproof, proof_l_com_y)))
+        mstore(add(_mPtr, 0x40), mload(add(aproof, proof_r_com_x)))
+        mstore(add(_mPtr, 0x60), mload(add(aproof, proof_r_com_y)))
+        mstore(add(_mPtr, 0x80), mload(add(aproof, proof_o_com_x)))
+        mstore(add(_mPtr, 0xa0), mload(add(aproof, proof_o_com_y)))
+        // pop(staticcall(sub(gas(), 2000), 0x2, add(mPtr, 0x1b), 0x365, mPtr, 0x20)) //0x1b -> 000.."gamma"
+
+        let size := add(0x2c5, mul(mload(pub_inputs), 0x20)) // 0x2c5 = 22*32+5
+        size := add(size, mul(vk_nb_commitments_commit_api, 0x40))
+        pop(staticcall(sub(gas(), 2000), 0x2, add(mPtr, 0x1b), size, mPtr, 0x20)) //0x1b -> 000.."gamma"
+      }
+
+      function derive_beta(aproof, prev_challenge){
+        let mPtr := mload(0x40)
+        // beta
+        mstore(mPtr, 0x62657461) // "beta"
+        mstore(add(mPtr, 0x20), prev_challenge)
+        pop(staticcall(sub(gas(), 2000), 0x2, add(mPtr, 0x1c), 0x24, mPtr, 0x20)) //0x1b -> 000.."gamma"
+      }
+
+      function derive_alpha(aproof, prev_challenge){
+        let mPtr := mload(0x40)
+        // alpha
+        mstore(mPtr, 0x616C706861) // "alpha"
+        mstore(add(mPtr, 0x20), prev_challenge)
+        mstore(add(mPtr, 0x40), mload(add(aproof, proof_grand_product_commitment_x)))
+        mstore(add(mPtr, 0x60), mload(add(aproof, proof_grand_product_commitment_y)))
+        pop(staticcall(sub(gas(), 2000), 0x2, add(mPtr, 0x1b), 0x65, mPtr, 0x20)) //0x1b -> 000.."gamma"
+      }
+
+      function derive_zeta(aproof, prev_challenge) {
+        let mPtr := mload(0x40)
+        // zeta
+        mstore(mPtr, 0x7a657461) // "zeta"
+        mstore(add(mPtr, 0x20), prev_challenge)
+        mstore(add(mPtr, 0x40), mload(add(aproof, proof_h_0_x)))
+        mstore(add(mPtr, 0x60), mload(add(aproof, proof_h_0_y)))
+        mstore(add(mPtr, 0x80), mload(add(aproof, proof_h_1_x)))
+        mstore(add(mPtr, 0xa0), mload(add(aproof, proof_h_1_y)))
+        mstore(add(mPtr, 0xc0), mload(add(aproof, proof_h_2_x)))
+        mstore(add(mPtr, 0xe0), mload(add(aproof, proof_h_2_y)))
+        pop(staticcall(sub(gas(), 2000), 0x2, add(mPtr, 0x1c), 0xe4, mPtr, 0x20))
+      }
     }
 
+    return (gamma, beta, alpha, zeta);
+  }
 
-    function deserialize_proof(
+  function load_wire_commitments_commit_api(uint256[] memory wire_commitments, bytes memory proof)
+  internal {
+    assembly {
+      let w := add(wire_commitments, 0x20)
+      let p := add(proof, proof_openings_selector_commit_api_at_zeta)
+      p := add(p, mul(vk_nb_commitments_commit_api, 0x20))
+      for {let i:=0} lt(i, mul(vk_nb_commitments_commit_api,2)) {i:=add(i,1)}
+      {
+        mstore(w, mload(p))
+        w := add(w,0x20)
+        p := add(p,0x20)
+        mstore(w, mload(p))
+        w := add(w,0x20)
+        p := add(p,0x20)
+      }
+    }
+  }
+
+  function compute_ith_lagrange_at_z(uint256 zeta, uint256 i) 
+  internal returns (uint256) {
+
+    uint256 res;
+    assembly {
+
+      // _n^_i [r]
+      function pow_local(x, e)->result {
+          let mPtr := mload(0x40)
+          mstore(mPtr, 0x20)
+          mstore(add(mPtr, 0x20), 0x20)
+          mstore(add(mPtr, 0x40), 0x20)
+          mstore(add(mPtr, 0x60), x)
+          mstore(add(mPtr, 0x80), e)
+          mstore(add(mPtr, 0xa0), r_mod)
+          pop(staticcall(sub(gas(), 2000),0x05,mPtr,0xc0,0x00,0x20))
+          result := mload(0x00)
+      }
+
+      let w := pow_local(vk_omega,i) // w**i
+      i := addmod(zeta, sub(r_mod, w), r_mod) // z-w**i
+      zeta := pow_local(zeta, vk_domain_size) // z**n
+      zeta := addmod(zeta, sub(r_mod, 1), r_mod) // z**n-1
+      w := mulmod(w, vk_inv_domain_size, r_mod) // w**i/n
+      i := pow_local(i, sub(r_mod,2)) // (z-w**i)**-1
+      w := mulmod(w, i, r_mod) // w**i/n*(z-w)**-1
+      res := mulmod(w, zeta, r_mod)
+    }
+    
+    return res;
+  }
+
+  function compute_pi(
+        bytes memory proof,
         uint256[] memory public_inputs,
-        uint256[] memory serialized_proof
-    ) internal pure returns(Proof memory proof) {
-        require(serialized_proof.length == SERIALIZED_PROOF_LENGTH);
-        proof.input_values = new uint256[](public_inputs.length);
-        for (uint256 i = 0; i < public_inputs.length; i++) {
-            proof.input_values[i] = public_inputs[i];
+        uint256 zeta
+    ) internal returns (uint256) {
+
+      // evaluation of Z=Xⁿ⁻¹ at ζ
+      // uint256 zeta_power_n_minus_one = Fr.pow(zeta, vk_domain_size);
+      // zeta_power_n_minus_one = Fr.sub(zeta_power_n_minus_one, 1);
+      uint256 zeta_power_n_minus_one;
+
+      uint256 pi;
+
+      assembly {
+        
+        sum_pi_wo_api_commit(add(public_inputs,0x20), mload(public_inputs), zeta)
+        pi := mload(mload(0x40))
+
+        function sum_pi_wo_api_commit(ins, n, z) {
+          let li := mload(0x40)
+          batch_compute_lagranges_at_z(z, n, li)
+          let res := 0
+          let tmp := 0
+          for {let i:=0} lt(i,n) {i:=add(i,1)}
+          {
+            tmp := mulmod(mload(li), mload(ins), r_mod)
+            res := addmod(res, tmp, r_mod)
+            li := add(li, 0x20)
+            ins := add(ins, 0x20)
+          }
+          mstore(mload(0x40), res)
         }
 
-        uint256 j = 0;
-        for (uint256 i = 0; i < STATE_WIDTH; i++) {
-            proof.wire_commitments[i] = PairingsBn254.new_g1_checked(
-                serialized_proof[j],
-                serialized_proof[j+1]
-            );
+        // mPtr <- [L_0(z), .., L_{n-1}(z)]
+        function batch_compute_lagranges_at_z(z, n, mPtr) {
+          let zn := addmod(pow(z, vk_domain_size, mPtr), sub(r_mod, 1), r_mod)
+          zn := mulmod(zn, vk_inv_domain_size, r_mod)
+          let _w := 1
+          let _mPtr := mPtr
+          for {let i:=0} lt(i,n) {i:=add(i,1)}
+          {
+            mstore(_mPtr, addmod(z,sub(r_mod, _w), r_mod))
+            _w := mulmod(_w, vk_omega, r_mod)
+            _mPtr := add(_mPtr, 0x20)
+          }
+          batch_invert(mPtr, n, _mPtr)
+          _mPtr := mPtr
+          _w := 1
+          for {let i:=0} lt(i,n) {i:=add(i,1)}
+          {
+            mstore(_mPtr, mulmod(mulmod(mload(_mPtr), zn , r_mod), _w, r_mod))
+            _mPtr := add(_mPtr, 0x20)
+            _w := mulmod(_w, vk_omega, r_mod)
+          }
+        } 
 
-            j += 2;
+        function batch_invert(ins, nb_ins, mPtr) {
+          mstore(mPtr, 1)
+          let offset := 0
+          for {let i:=0} lt(i, nb_ins) {i:=add(i,1)}
+          {
+            let prev := mload(add(mPtr, offset))
+            let cur := mload(add(ins, offset))
+            cur := mulmod(prev, cur, r_mod)
+            offset := add(offset, 0x20)
+            mstore(add(mPtr, offset), cur)
+          }
+          ins := add(ins, sub(offset, 0x20))
+          mPtr := add(mPtr, offset)
+          let inv := pow(mload(mPtr), sub(r_mod,2), add(mPtr, 0x20))
+          for {let i:=0} lt(i, nb_ins) {i:=add(i,1)}
+          {
+            mPtr := sub(mPtr, 0x20)
+            let tmp := mload(ins)
+            let cur := mulmod(inv, mload(mPtr), r_mod)
+            mstore(ins, cur)
+            inv := mulmod(inv, tmp, r_mod)
+            ins := sub(ins, 0x20)
+          }
         }
 
-        proof.grand_product_commitment = PairingsBn254.new_g1_checked(
-            serialized_proof[j],
-            serialized_proof[j+1]
-        );
-        j += 2;
-
-        for (uint256 i = 0; i < STATE_WIDTH; i++) {
-            proof.quotient_poly_commitments[i] = PairingsBn254.new_g1_checked(
-                serialized_proof[j],
-                serialized_proof[j+1]
-            );
-
-            j += 2;
+        // res <- x^e mod r
+        function pow(x, e, mPtr)->res {
+          mstore(mPtr, 0x20)
+          mstore(add(mPtr, 0x20), 0x20)
+          mstore(add(mPtr, 0x40), 0x20)
+          mstore(add(mPtr, 0x60), x)
+          mstore(add(mPtr, 0x80), e)
+          mstore(add(mPtr, 0xa0), r_mod)
+          pop(staticcall(sub(gas(), 2000),0x05,mPtr,0xc0,mPtr,0x20))
+          res := mload(mPtr)
         }
 
-        for (uint256 i = 0; i < STATE_WIDTH; i++) {
-            proof.wire_values_at_zeta[i] = PairingsBn254.new_fr(
-                serialized_proof[j]
-            );
+        zeta_power_n_minus_one := pow(zeta, vk_domain_size, mload(0x40))
+        zeta_power_n_minus_one := addmod(zeta_power_n_minus_one, sub(r_mod, 1), r_mod)
+      }
 
-            j += 1;
-        }
+      {{ if (gt (len .CommitmentConstraintIndexes) 0 )}}
+      uint256[] memory commitment_indices = new uint256[](vk_nb_commitments_commit_api);
+      load_vk_commitments_indices_commit_api(commitment_indices);
 
-        proof.grand_product_at_zeta_omega = PairingsBn254.new_fr(
-            serialized_proof[j]
-        );
+      uint256[] memory wire_committed_commitments;
+      wire_committed_commitments = new uint256[](2*vk_nb_commitments_commit_api);
 
-        j += 1;
+      load_wire_commitments_commit_api(wire_committed_commitments, proof);
 
-        proof.quotient_polynomial_at_zeta = PairingsBn254.new_fr(
-            serialized_proof[j]
-        );
-
-        j += 1;
-
-        proof.linearization_polynomial_at_zeta = PairingsBn254.new_fr(
-            serialized_proof[j]
-        );
-
-        j += 1;
-
-        for (uint256 i = 0; i < proof.permutation_polynomials_at_zeta.length; i++) {
-            proof.permutation_polynomials_at_zeta[i] = PairingsBn254.new_fr(
-                serialized_proof[j]
-            );
-
-            j += 1;
-        }
-
-        proof.opening_at_zeta_proof = PairingsBn254.new_g1_checked(
-            serialized_proof[j],
-            serialized_proof[j+1]
-        );
-        j += 2;
-
-        proof.opening_at_zeta_omega_proof = PairingsBn254.new_g1_checked(
-            serialized_proof[j],
-            serialized_proof[j+1]
-        );
+      for (uint256 i=0; i<vk_nb_commitments_commit_api; i++){
+          
+          uint256 hash_res = Utils.hash_fr(wire_committed_commitments[2*i], wire_committed_commitments[2*i+1]);
+          uint256 a = compute_ith_lagrange_at_z(zeta, commitment_indices[i]+public_inputs.length);
+          assembly {
+            a := mulmod(hash_res, a, r_mod)
+            pi := addmod(pi, a, r_mod)
+          }
+      }
+      {{ end }}
+      
+      return pi;
     }
 
-    function verify_serialized_proof(
-        uint256[] memory public_inputs,
-        uint256[] memory serialized_proof
-    ) public view returns (bool) {
-        VerificationKey memory vk = get_verification_key();
-        require(vk.num_inputs == public_inputs.length);
-        Proof memory proof = deserialize_proof(public_inputs, serialized_proof);
-        bool valid = verify(proof, vk);
-        return valid;
+  function Verify(bytes memory proof, uint256[] memory public_inputs) 
+  internal returns(bool) {
+
+    uint256 gamma;
+    uint256 beta;
+    uint256 alpha;
+    uint256 zeta;
+
+    (gamma, beta, alpha, zeta) = derive_gamma_beta_alpha_zeta(proof, public_inputs);
+
+    uint256 pi = compute_pi(proof, public_inputs, zeta);
+
+    uint256 check;
+
+    bool success = false;
+    // uint256 success;
+
+    assembly {
+
+      let mem := mload(0x40)
+      mstore(add(mem, state_alpha), alpha)
+      mstore(add(mem, state_gamma), gamma)
+      mstore(add(mem, state_zeta), zeta)
+      mstore(add(mem, state_beta), beta)
+      mstore(add(mem, state_pi), pi)
+
+      compute_alpha_square_lagrange()
+      verify_quotient_poly_eval_at_zeta(proof)
+      fold_h(proof)
+      compute_commitment_linearised_polynomial(proof)
+      compute_gamma_kzg(proof)
+      fold_state(proof)
+      batch_verify_multi_points(proof)
+
+      success := mload(add(mem, state_success))
+      
+      check := mload(add(mem, state_check_var))
+
+      function compute_alpha_square_lagrange() {   
+        let state := mload(0x40)
+        let mPtr := add(mload(0x40), state_last_mem)
+
+        // zeta**n - 1
+        let res := pow(mload(add(state, state_zeta)), vk_domain_size, mPtr)
+        res := addmod(res, sub(r_mod,1), r_mod)
+        mstore(add(state, state_zeta_power_n_minus_one), res)
+
+        // let res := mload(add(state, state_zeta_power_n_minus_one))
+        let den := addmod(mload(add(state, state_zeta)), sub(r_mod, 1), r_mod)
+        den := pow(den, sub(r_mod, 2), mPtr)
+        den := mulmod(den, vk_inv_domain_size, r_mod)
+        res := mulmod(den, res, r_mod)
+
+        let l_alpha := mload(add(state, state_alpha))
+        res := mulmod(res, l_alpha, r_mod)
+        res := mulmod(res, l_alpha, r_mod)
+        mstore(add(state, state_alpha_square_lagrange), res)
+      }
+
+      function batch_verify_multi_points(aproof) {
+
+        let state := mload(0x40)
+        let mPtr := add(state, state_last_mem)
+
+        // here the random is not a challenge, hence no need to use Fiat Shamir, we just
+        // need an unpredictible result.
+        let random := mod(keccak256(state, 0x20), r_mod)
+
+        let folded_quotients := mPtr
+        mPtr := add(folded_quotients, 0x40)
+        mstore(folded_quotients, mload(add(aproof, proof_batch_opening_at_zeta_x)))
+        mstore(add(folded_quotients, 0x20), mload(add(aproof, proof_batch_opening_at_zeta_y)))
+        point_acc_mul(folded_quotients, add(aproof, proof_opening_at_zeta_omega_x), random, mPtr)
+
+        let folded_digests := add(state, state_folded_digests_x)
+        point_acc_mul(folded_digests, add(aproof, proof_grand_product_commitment_x), random, mPtr)
+
+        let folded_evals := add(state, state_folded_claimed_values)
+        fr_acc_mul(folded_evals, add(aproof, proof_grand_product_at_zeta_omega), random)
+
+        let folded_evals_commit := mPtr
+        mPtr := add(folded_evals_commit, 0x40)
+        mstore(folded_evals_commit, 0x1)
+        mstore(add(folded_evals_commit, 0x20), 0x2)
+        mstore(add(folded_evals_commit, 0x40), mload(folded_evals))
+        pop(staticcall(sub(gas(), 2000),7,folded_evals_commit,0x60,folded_evals_commit,0x40))
+
+        let folded_evals_commit_y := add(folded_evals_commit, 0x20)
+        mstore(folded_evals_commit_y, sub(p_mod, mload(folded_evals_commit_y)))
+        point_add(folded_digests, folded_digests, folded_evals_commit, mPtr)
+
+        let folded_points_quotients := mPtr
+        mPtr := add(mPtr, 0x40)
+        point_mul(folded_points_quotients, add(aproof, proof_batch_opening_at_zeta_x), mload(add(state, state_zeta)), mPtr)
+        let zeta_omega := mulmod(mload(add(state, state_zeta)), vk_omega, r_mod)
+        random := mulmod(random, zeta_omega, r_mod)
+        point_acc_mul(folded_points_quotients, add(aproof, proof_opening_at_zeta_omega_x), random, mPtr)
+
+        point_add(folded_digests, folded_digests, folded_points_quotients, mPtr)
+
+        let folded_quotients_y := add(folded_quotients, 0x20)
+        mstore(folded_quotients_y, sub(p_mod, mload(folded_quotients_y)))
+
+        mstore(mPtr, mload(folded_digests))
+        mstore(add(mPtr, 0x20), mload(add(folded_digests, 0x20)))
+        mstore(add(mPtr, 0x40), g2_srs_0_x_0) // the 4 lines are the canonical G2 point on BN254
+        mstore(add(mPtr, 0x60), g2_srs_0_x_1)
+        mstore(add(mPtr, 0x80), g2_srs_0_y_0)
+        mstore(add(mPtr, 0xa0), g2_srs_0_y_1)
+        mstore(add(mPtr, 0xc0), mload(folded_quotients))
+        mstore(add(mPtr, 0xe0), mload(add(folded_quotients, 0x20)))
+        mstore(add(mPtr, 0x100), g2_srs_1_x_0)
+        mstore(add(mPtr, 0x120), g2_srs_1_x_1)
+        mstore(add(mPtr, 0x140), g2_srs_1_y_0)
+        mstore(add(mPtr, 0x160), g2_srs_1_y_1)
+        let l_success := staticcall(sub(gas(), 2000),8,mPtr,0x180,0x00,0x20)
+        // l_success := true
+        mstore(add(state, state_success), and(l_success,mload(add(state, state_success))))
+      }
+
+      // at this stage the state of mPtr is the same as in compute_gamma
+      function fold_state(aproof) {
+        
+        let state := mload(0x40)
+        let mPtr := add(mload(0x40), state_last_mem)
+
+        let l_gamma_kzg := mload(add(state, state_gamma_kzg))
+        let acc_gamma := l_gamma_kzg
+
+        let offset := add(0x200, mul(vk_nb_commitments_commit_api, 0x40)) // 0x40 = 2*0x20
+        let mPtrOffset := add(mPtr, offset)
+
+        mstore(add(state, state_folded_digests_x), mload(add(mPtr,0x40)))
+        mstore(add(state, state_folded_digests_y), mload(add(mPtr,0x60)))
+        mstore(add(state, state_folded_claimed_values), mload(add(aproof, proof_quotient_polynomial_at_zeta)))
+
+        point_acc_mul(add(state, state_folded_digests_x), add(mPtr,0x80), acc_gamma, mPtrOffset)
+        fr_acc_mul(add(state, state_folded_claimed_values), add(aproof, proof_linearised_polynomial_at_zeta), acc_gamma)
+        mstore(add(state, state_check_var), acc_gamma)
+        
+        acc_gamma := mulmod(acc_gamma, l_gamma_kzg, r_mod)
+        point_acc_mul(add(state, state_folded_digests_x), add(mPtr,0xc0), acc_gamma, mPtrOffset)
+        fr_acc_mul(add(state, state_folded_claimed_values), add(aproof, proof_l_at_zeta), acc_gamma)
+        
+        acc_gamma := mulmod(acc_gamma, l_gamma_kzg, r_mod)
+        point_acc_mul(add(state, state_folded_digests_x), add(mPtr,0x100), acc_gamma, add(mPtr, offset))
+        fr_acc_mul(add(state, state_folded_claimed_values), add(aproof, proof_r_at_zeta), acc_gamma)
+
+        acc_gamma := mulmod(acc_gamma, l_gamma_kzg, r_mod)
+        point_acc_mul(add(state, state_folded_digests_x), add(mPtr,0x140), acc_gamma, add(mPtr, offset))
+        fr_acc_mul(add(state, state_folded_claimed_values), add(aproof, proof_o_at_zeta), acc_gamma)
+        
+        acc_gamma := mulmod(acc_gamma, l_gamma_kzg, r_mod)
+        point_acc_mul(add(state, state_folded_digests_x), add(mPtr,0x180), acc_gamma, add(mPtr, offset))
+        fr_acc_mul(add(state, state_folded_claimed_values), add(aproof, proof_s1_at_zeta), acc_gamma)
+        
+        acc_gamma := mulmod(acc_gamma, l_gamma_kzg, r_mod)
+        point_acc_mul(add(state, state_folded_digests_x), add(mPtr,0x1c0), acc_gamma, add(mPtr, offset))
+        fr_acc_mul(add(state, state_folded_claimed_values), add(aproof, proof_s2_at_zeta), acc_gamma)
+        
+        let poscaz := add(aproof, proof_openings_selector_commit_api_at_zeta)
+        let opca := add(mPtr, 0x200) // offset_proof_commits_api
+        for {let i := 0} lt(i, vk_nb_commitments_commit_api) {i:=add(i,1)}
+        {
+          acc_gamma := mulmod(acc_gamma, l_gamma_kzg, r_mod)
+          point_acc_mul(add(state, state_folded_digests_x), opca, acc_gamma, add(mPtr, offset))
+          fr_acc_mul(add(state, state_folded_claimed_values), poscaz, acc_gamma)
+          poscaz := add(poscaz, 0x20)
+          opca := add(opca, 0x40)
+        }
+
+      }
+
+      function compute_gamma_kzg(aproof) {
+
+        let state := mload(0x40)
+        let mPtr := add(mload(0x40), state_last_mem)
+        mstore(mPtr, 0x67616d6d61) // "gamma"
+        mstore(add(mPtr, 0x20), mload(add(state, state_zeta)))
+        mstore(add(mPtr,0x40), mload(add(state, state_folded_h_x)))
+        mstore(add(mPtr,0x60), mload(add(state, state_folded_h_y)))
+        mstore(add(mPtr,0x80), mload(add(state, state_linearised_polynomial_x)))
+        mstore(add(mPtr,0xa0), mload(add(state, state_linearised_polynomial_y)))
+        mstore(add(mPtr,0xc0), mload(add(aproof, proof_l_com_x)))
+        mstore(add(mPtr,0xe0), mload(add(aproof, proof_l_com_y)))
+        mstore(add(mPtr,0x100), mload(add(aproof, proof_r_com_x)))
+        mstore(add(mPtr,0x120), mload(add(aproof, proof_r_com_y)))
+        mstore(add(mPtr,0x140), mload(add(aproof, proof_o_com_x)))
+        mstore(add(mPtr,0x160), mload(add(aproof, proof_o_com_y)))
+        mstore(add(mPtr,0x180), vk_s1_com_x)
+        mstore(add(mPtr,0x1a0), vk_s1_com_y)
+        mstore(add(mPtr,0x1c0), vk_s2_com_x)
+        mstore(add(mPtr,0x1e0), vk_s2_com_y)
+        
+        let offset := 0x200
+        {{ range $index, $element := .CommitmentConstraintIndexes }}
+        mstore(add(mPtr,offset), vk_selector_commitments_commit_api_{{ $index }}_x)
+        mstore(add(mPtr,add(offset, 0x20)), vk_selector_commitments_commit_api_{{ $index }}_y)
+        offset := add(offset, 0x40)
+        {{ end }}
+
+        mstore(add(mPtr, offset), mload(add(aproof, proof_quotient_polynomial_at_zeta)))
+        mstore(add(mPtr, add(offset, 0x20)), mload(add(aproof, proof_linearised_polynomial_at_zeta)))
+        mstore(add(mPtr, add(offset, 0x40)), mload(add(aproof, proof_l_at_zeta)))
+        mstore(add(mPtr, add(offset, 0x60)), mload(add(aproof, proof_r_at_zeta)))
+        mstore(add(mPtr, add(offset, 0x80)), mload(add(aproof, proof_o_at_zeta)))
+        mstore(add(mPtr, add(offset, 0xa0)), mload(add(aproof, proof_s1_at_zeta)))
+        mstore(add(mPtr, add(offset, 0xc0)), mload(add(aproof, proof_s2_at_zeta)))
+
+        {{ if (gt (len .CommitmentConstraintIndexes) 0 )}}
+        let _mPtr := add(mPtr, add(offset, 0xe0))
+        let _poscaz := add(aproof, proof_openings_selector_commit_api_at_zeta)
+        for {let i:=0} lt(i, vk_nb_commitments_commit_api) {i:=add(i,1)}
+        {
+          mstore(_mPtr, mload(_poscaz))
+          _poscaz := add(_poscaz, 0x20)
+          _mPtr := add(_mPtr, 0x20)
+        }
+        {{ end }}
+
+        let start_input := 0x1b // 00.."gamma"
+        let size_input := add(0x16, mul(vk_nb_commitments_commit_api,3)) // number of 32bytes elmts = 0x16 (zeta+2*7+7 for the digests+openings) + 2*vk_nb_commitments_commit_api (for the commitments of the selectors) + vk_nb_commitments_commit_api (for the openings of the selectors)
+        size_input := add(0x5, mul(size_input, 0x20)) // size in bytes: 15*32 bytes + 5 bytes for gamma
+        pop(staticcall(sub(gas(), 2000), 0x2, add(mPtr,start_input), size_input, add(state, state_gamma_kzg), 0x20))
+        mstore(add(state, state_gamma_kzg), mod(mload(add(state, state_gamma_kzg)), r_mod))
+      }
+
+      function compute_commitment_linearised_polynomial_ec(aproof, s1, s2) {
+
+        let state := mload(0x40)
+        let mPtr := add(mload(0x40), state_last_mem)
+
+        mstore(mPtr, vk_ql_com_x)
+        mstore(add(mPtr,0x20), vk_ql_com_y)
+        point_mul(add(state, state_linearised_polynomial_x), mPtr, mload(add(aproof, proof_l_at_zeta)), add(mPtr,0x40))
+
+        mstore(mPtr, vk_qr_com_x)
+        mstore(add(mPtr,0x20), vk_qr_com_y)
+        point_acc_mul(add(state, state_linearised_polynomial_x),mPtr,mload(add(aproof, proof_r_at_zeta)),add(mPtr,0x40))
+        
+        let rl := mulmod(mload(add(aproof, proof_l_at_zeta)), mload(add(aproof, proof_r_at_zeta)), r_mod)
+        mstore(mPtr, vk_qm_com_x)
+        mstore(add(mPtr,0x20), vk_qm_com_y)
+        point_acc_mul(add(state, state_linearised_polynomial_x),mPtr,rl,add(mPtr,0x40))
+        
+        mstore(mPtr, vk_qo_com_x)
+        mstore(add(mPtr,0x20), vk_qo_com_y)
+        point_acc_mul(add(state, state_linearised_polynomial_x),mPtr,mload(add(aproof, proof_o_at_zeta)),add(mPtr,0x40))
+        
+        mstore(mPtr, vk_qk_com_x)
+        mstore(add(mPtr, 0x20), vk_qk_com_y)
+        point_add(add(state, state_linearised_polynomial_x),add(state, state_linearised_polynomial_x),mPtr,add(mPtr, 0x40))
+
+        let commits_api_at_zeta := add(aproof, proof_openings_selector_commit_api_at_zeta)
+        let commits_api := add(aproof, add(proof_openings_selector_commit_api_at_zeta, mul(vk_nb_commitments_commit_api, 0x20)))
+        for {let i:=0} lt(i, vk_nb_commitments_commit_api) {i:=add(i,1)}
+        {
+          mstore(mPtr, mload(commits_api))
+          mstore(add(mPtr, 0x20), mload(add(commits_api, 0x20)))
+          point_acc_mul(add(state, state_linearised_polynomial_x),mPtr,mload(commits_api_at_zeta),add(mPtr,0x40))
+          commits_api_at_zeta := add(commits_api_at_zeta, 0x20)
+          commits_api := add(commits_api, 0x40)
+        }
+
+        mstore(mPtr, vk_s3_com_x)
+        mstore(add(mPtr, 0x20), vk_s3_com_y)
+        point_acc_mul(add(state, state_linearised_polynomial_x), mPtr, s1, add(mPtr, 0x40))
+
+        mstore(mPtr, mload(add(aproof, proof_grand_product_commitment_x)))
+        mstore(add(mPtr, 0x20), mload(add(aproof, proof_grand_product_commitment_y)))
+        point_acc_mul(add(state, state_linearised_polynomial_x), mPtr, s2, add(mPtr, 0x40))
+
+      }
+
+      function compute_commitment_linearised_polynomial(aproof) {
+        
+        let state := mload(0x40)
+        let l_beta := mload(add(state, state_beta))
+        let l_gamma := mload(add(state, state_gamma))
+        let l_zeta := mload(add(state, state_zeta))
+        let l_alpha := mload(add(state, state_alpha))
+
+        let u := mulmod(mload(add(aproof,proof_grand_product_at_zeta_omega)), l_beta, r_mod)
+        let v := mulmod(l_beta, mload(add(aproof, proof_s1_at_zeta)), r_mod)
+        v := addmod(v, mload(add(aproof, proof_l_at_zeta)), r_mod)
+        v := addmod(v, l_gamma, r_mod)
+
+        let w := mulmod(l_beta, mload(add(aproof, proof_s2_at_zeta)), r_mod)
+        w := addmod(w, mload(add(aproof, proof_r_at_zeta)), r_mod)
+        w := addmod(w, l_gamma, r_mod)
+
+        let s1 := mulmod(u, v, r_mod)
+        s1 := mulmod(s1, w, r_mod)
+        s1 := mulmod(s1, l_alpha, r_mod)
+
+        let coset_square := mulmod(vk_coset_shift, vk_coset_shift, r_mod)
+        let betazeta := mulmod(l_beta, l_zeta, r_mod)
+        u := addmod(betazeta, mload(add(aproof, proof_l_at_zeta)), r_mod)
+        u := addmod(u, l_gamma, r_mod)
+
+        v := mulmod(betazeta, vk_coset_shift, r_mod)
+        v := addmod(v, mload(add(aproof, proof_r_at_zeta)), r_mod)
+        v := addmod(v, l_gamma, r_mod)
+
+        w := mulmod(betazeta, coset_square, r_mod)
+        w := addmod(w, mload(add(aproof, proof_o_at_zeta)), r_mod)
+        w := addmod(w, l_gamma, r_mod)
+
+        let s2 := mulmod(u, v, r_mod)
+        s2 := mulmod(s2, w, r_mod)
+        s2 := sub(r_mod, s2)
+        s2 := mulmod(s2, l_alpha, r_mod)
+        s2 := addmod(s2, mload(add(state, state_alpha_square_lagrange)), r_mod)
+
+        compute_commitment_linearised_polynomial_ec(aproof, s1, s2)
+      }
+
+      function fold_h(aproof) {
+        let state := mload(0x40)
+        let n_plus_two := add(vk_domain_size, 2)
+        let mPtr := add(mload(0x40), state_last_mem)
+        let zeta_power_n_plus_two := pow(mload(add(state, state_zeta)), n_plus_two, mPtr)
+        point_mul(add(state, state_folded_h_x), add(aproof, proof_h_2_x), zeta_power_n_plus_two, mPtr)
+        point_add(add(state, state_folded_h_x), add(state, state_folded_h_x), add(aproof, proof_h_1_x), mPtr)
+        point_mul(add(state, state_folded_h_x), add(state, state_folded_h_x), zeta_power_n_plus_two, mPtr)
+        point_add(add(state, state_folded_h_x), add(state, state_folded_h_x), add(aproof, proof_h_0_x), mPtr)
+      }
+
+      function verify_quotient_poly_eval_at_zeta(aproof) {
+
+        let state := mload(0x40)
+
+        // (l(ζ)+β*s1(ζ)+γ)
+        let s1 := add(mload(0x40), state_last_mem)
+        mstore(s1, mulmod(mload(add(aproof,proof_s1_at_zeta)),mload(add(state, state_beta)), r_mod))
+        mstore(s1, addmod(mload(s1), mload(add(state, state_gamma)), r_mod))
+        mstore(s1, addmod(mload(s1), mload(add(aproof, proof_l_at_zeta)), r_mod))
+
+        // (r(ζ)+β*s2(ζ)+γ)
+        let s2 := add(s1,0x20)
+        mstore(s2, mulmod(mload(add(aproof,proof_s2_at_zeta)),mload(add(state, state_beta)), r_mod))
+        mstore(s2, addmod(mload(s2), mload(add(state, state_gamma)), r_mod))
+        mstore(s2, addmod(mload(s2), mload(add(aproof, proof_r_at_zeta)), r_mod))
+        // _s2 := mload(s2)
+
+        // (o(ζ)+γ)
+        let o := add(s1,0x40)
+        mstore(o, addmod(mload(add(aproof,proof_o_at_zeta)), mload(add(state, state_gamma)), r_mod))
+
+        //  α*(Z(μζ))*(l(ζ)+β*s1(ζ)+γ)*(r(ζ)+β*s2(ζ)+γ)*(o(ζ)+γ)
+        mstore(s1, mulmod(mload(s1), mload(s2), r_mod))
+        mstore(s1, mulmod(mload(s1), mload(o), r_mod))
+        mstore(s1, mulmod(mload(s1), mload(add(state, state_alpha)), r_mod))
+        mstore(s1, mulmod(mload(s1), mload(add(aproof, proof_grand_product_at_zeta_omega)), r_mod))
+
+        let computed_quotient := add(s1,0x60)
+
+        // linearizedpolynomial + pi(zeta)
+        mstore(computed_quotient, addmod(mload(add(aproof, proof_linearised_polynomial_at_zeta)), mload(add(state, state_pi)), r_mod))
+        mstore(computed_quotient, addmod(mload(computed_quotient), mload(s1), r_mod))
+        mstore(computed_quotient, addmod(mload(computed_quotient), sub(r_mod,mload(add(state, state_alpha_square_lagrange))), r_mod))
+        mstore(s2, mulmod(mload(add(aproof,proof_quotient_polynomial_at_zeta)), mload(add(state, state_zeta_power_n_minus_one)), r_mod))
+
+        mstore(add(state, state_success),eq(mload(computed_quotient), mload(s2)))
+      }
+
+      function point_add(dst, p, q, mPtr) {
+        // let mPtr := add(mload(0x40), state_last_mem)
+        let state := mload(0x40)
+        mstore(mPtr, mload(p))
+        mstore(add(mPtr, 0x20), mload(add(p, 0x20)))
+        mstore(add(mPtr, 0x40), mload(q))
+        mstore(add(mPtr, 0x60), mload(add(q, 0x20)))
+        let l_success := staticcall(sub(gas(), 2000),6,mPtr,0x80,dst,0x40)
+        mstore(add(state, state_success), and(l_success,mload(add(state, state_success))))
+      }
+
+      // dst <- [s]src
+      function point_mul(dst,src,s, mPtr) {
+        // let mPtr := add(mload(0x40), state_last_mem)
+        let state := mload(0x40)
+        mstore(mPtr,mload(src))
+        mstore(add(mPtr,0x20),mload(add(src,0x20)))
+        mstore(add(mPtr,0x40),s)
+        let l_success := staticcall(sub(gas(), 2000),7,mPtr,0x60,dst,0x40)
+        mstore(add(state, state_success), and(l_success,mload(add(state, state_success))))
+      }
+
+      // dst <- dst + [s]src (Elliptic curve)
+      function point_acc_mul(dst,src,s, mPtr) {
+        let state := mload(0x40)
+        mstore(mPtr,mload(src))
+        mstore(add(mPtr,0x20),mload(add(src,0x20)))
+        mstore(add(mPtr,0x40),s)
+        let l_success := staticcall(sub(gas(), 2000),7,mPtr,0x60,mPtr,0x40)
+        mstore(add(mPtr,0x40),mload(dst))
+        mstore(add(mPtr,0x60),mload(add(dst,0x20)))
+        l_success := and(l_success, staticcall(sub(gas(), 2000),6,mPtr,0x80,dst, 0x40))
+        mstore(add(state, state_success), and(l_success,mload(add(state, state_success))))
+      }
+
+      // dst <- dst + src (Fr) dst,src are addresses, s is a value
+      function fr_acc_mul(dst, src, s) {
+        let tmp :=  mulmod(mload(src), s, r_mod)
+        mstore(dst, addmod(mload(dst), tmp, r_mod))
+      }
+
+      // dst <- x ** e mod r (x, e are values, not pointers)
+      function pow(x, e, mPtr)->res {
+        mstore(mPtr, 0x20)
+        mstore(add(mPtr, 0x20), 0x20)
+        mstore(add(mPtr, 0x40), 0x20)
+        mstore(add(mPtr, 0x60), x)
+        mstore(add(mPtr, 0x80), e)
+        mstore(add(mPtr, 0xa0), r_mod)
+        pop(staticcall(sub(gas(), 2000),0x05,mPtr,0xc0,mPtr,0x20))
+        res := mload(mPtr)
+      }
     }
+
+    return success;
+
+  }
+
 }
 `
