+++ backend/plonk/bn254/verify.go
@@ -24,6 +24,7 @@ import (
 	"time"
 
 	"github.com/consensys/gnark-crypto/ecc/bn254/fr"
+	"github.com/consensys/gnark-crypto/ecc/bn254/fp"
 
 	"github.com/consensys/gnark-crypto/ecc/bn254/fr/kzg"
 
@@ -346,9 +347,28 @@ func deriveRandomness(fs *fiatshamir.Transcript, challenge string, points ...*cu
 //
 // Code has not been audited and is provided as-is, we make no guarantees or warranties to its safety and reliability.
 func (vk *VerifyingKey) ExportSolidity(w io.Writer) error {
-	tmpl, err := template.New("").Parse(solidityTemplate)
-	if err != nil {
-		return err
+	funcMap := template.FuncMap{
+		// The name "inc" is what the function will be called in the template text.
+		"inc": func(i int) int {
+			return i + 1
+		},
+		"frptr": func(x fr.Element) *fr.Element {
+			return &x
+		},
+		"fpptr": func(x fp.Element) *fp.Element {
+			return &x
+		},
+		"add": func(i, j int) int {
+			return i + j
+		},
 	}
-	return tmpl.Execute(w, vk)
+
+
+		t, err := template.New("t").Funcs(funcMap).Parse(tmplSolidityVerifier)
+		if err != nil {
+			return err
+		}
+		return t.Execute(w, vk)
+
+		
 }
